web/pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
#packages = [{ include = "src" }]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.pylint.master]
init-hook = "import sys; sys.path.append('./web/src')"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
